Fix refetching renamed repos
In certain cases, repos' have been renamed and then recreated with the same name and
a different id. When attempting to fetch the old name with original id we were
inserting a stub entry under the original id, which was overwriting the new repository
name with the old one. Fix.